move get_docker_user to backend_utils.py
